Fix compile bug
Old repository SHA1: 1820eabdda00fab1755a90e548501c1a862879d7

+++ source/src/core/scoring/cryst/XtalMLEnergy.cc
@@ -148,7 +148,7 @@ void XtalMLEnergy::eval_atom_derivative(
 	core::conformation::Residue const &rsd_i = pose.residue(resid);
 
 	if ( rsd_i.aa() == core::chemical::aa_vrt ) return;
-	if ( resid > dml_dx.size() ) return; // calling from outside ASU (since this is wholestructure energy it calls on everything)
+	if ( resid > (int)dml_dx.size() ) return; // calling from outside ASU (since this is wholestructure energy it calls on everything)
 
 	// look up derivative
 	numeric::xyzVector< core::Real > dCCdx = dml_dx[resid][atmid];
